requirements.txt
discord.py
python-dotenv
playwright
python-telegram-bot>=20.7